requirements.txt
nltk==3.5
tweepy==3.10.0
python-dotenv==0.15.0
numpy==1.20.1
pandas==1.2.4
